+++ apps/boruta_federation/lib/boruta_federation_web/controllers/openid_controller.ex
@@ -1,0 +1,11 @@
+defmodule BorutaFederationWeb.OpenidController do
+  use BorutaFederationWeb, :controller
+
+  alias BorutaFederation.FederationEntities
+
+  def well_known(conn, params) do
+    entity = FederationEntities.get_entity_by_id(params["entity_id"])
+
+    render(conn, "well_known.jwt", entity: entity)
+  end
+end